Popravljena metoda za ipis in lepsi zapis metode getKey v src/fields/Address.java. Popravljena metoda equals v src/fields/Cookie.java. Popravljena vidnost metod v vmesniku src/fields/Field.java. Odstanjevanje metod za FTP v src/fields/Method.java. Popravek konstruktorja zaraidi odstanitve FTP protokola v src/fields/ncsa/RequestLine.java. Popravek metode parseLine v src/parser/IISParser.java. Popravek metode parseLine v src/parser/NCSAParser.java. Popravek metode parseLine v src/parser/W3CParser.java. Dodani testi za testiranje iteratorja v test/datastruct/RadixTreeTest.java. Popravek testov za testiranje parsanja v test/logparser/TestParser.java.

+++ src/spletneseje/SpletneSeje.java
@@ -106,9 +106,7 @@ public class SpletneSeje {
 
 		//Ustvari novo izhodno datoteko
 		File outFile = new File(argsParser.getOutputFilePath());
-		if(argsParser.deleteOutputFile()) {
-			outFile.delete();
-		}
+		if(argsParser.deleteOutputFile()) outFile.delete();
 		outDataBase = SqlJetDb.open(outFile, true);
 
 		// TODO izdelaj podatkovno bazo oziroma, ce datoteka ze obstaja preveri ali uporablja pravilne tabele, če temu ni tako javi napako
